Turing Machine
The parenthesis expression is enclosed between the left and right limiters, indicated by the symbol A. Machine operation: first, the RNG moves to the right to the first right bracket, replaces it with the symbol X, goes to the state q1 and moves to the left to the nearest left bracket, replaces it with the symbol X, goes to the state q0, and the shuttle movement repeats. If the machine, while in state q1, reaches the left character A, it prints "0" and stops – the parenthesis is incorrect. If the machine, while in state q0, reaches the right symbol A without finding the right bracket anymore, it goes into the final state q2, associated with moving to the left, and looks at the sequence for the last time: is there an unpaired left bracket left. If a left bracket is encountered along the way, the machine prints "0" and stops. When the left character A is reached in the q2 state, the machine prints "1" and stops – the parenthesis expression is correct.