hash
Consider the following task: there is a data file that is transmitted over the Internet. After receiving this file, it is necessary to check whether the file was corrupted during transmission.
One of the ways to check: we consider the file as a sequence of bits, divide it into blocks of 32 bits and add the resulting numbers, discarding the overflow.
If the file starts with the following bits:
10100111 01011100 10010100 01111111 10010001 10001101 10010100 00010000 00011010 00001000 10010100 01111111 00010001 10001101 10010100 01111111
We take the first 32 bits:
10100111 01011100 10010100 01111111
And add the following 32 bits to them:
10010001 10001101 10010100 00010000
It turns out:
1 00111000 11101010 00101000 10001111
Discarding overflow:
00111000 11101010 00101000 10001111
This way you can sum up the entire file. As a result, we will get a number (a sequence of bits). This number is called the hash sum. It should be written to the end of the file. When receiving a file on another machine, it is necessary to recalculate the hash amount and compare the result.